requirements.txt
alembic==1.11.3
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
environs==9.5.0
exceptiongroup==1.1.3
fastapi==0.103.0
FastAPI-SQLAlchemy==0.2.1
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.1
psycopg2-binary==2.9.7
pydantic==1.10.12
pydantic_core==2.6.3
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
xlwings==0.30.11
